Refactor code and unit tests to allow for multiple EventBus implementations.
